change adapters to be self contained Objects
This allows for an adapter to be attached as a whole (versus the piece
by piece sub/unsub/get/set) which makes it easier to bind a while new
adapter. Since all getting/settings occurs through the adapter, this
will also facilitate model-less or view specific properties by
consolidating the interaction between the reactive instance and the
model.