fix event binding after 'each' has been bound
Previously, bindings for on-click, etc which occurred inside of an
'each' binding zone would be bound erroneously to the parent in addition
to being bound to each item in the iteration. This would case too many
events to fire.

The problem was caused by the binding layer not properly detecting that
the 'each' binding had essentially pruned away what was inside of it and
that those event bindings would be handled separately.

This commit changes how events are bound by doing a walk from the parent
node down into all of the children. Each level of the walk can be
stopped from traversing deeper by setting a .skip flag inside the
binding on the `this` object (which is the Binding itself). Upon seeing
a skip, the walk will no longer continue.